Refactor pojo to models
