+++ app/src/main/java/com/ss20/se2/monopoly/models/deeds/Card.java
@@ -1,4 +1,4 @@
-package com.ss20.se2.monopoly.pojo.models.deeds;
+package com.ss20.se2.monopoly.models.deeds;
 
 public abstract class Card extends GameTile{
