+++ app/src/main/java/com/ss20/se2/monopoly/network/OnLocalGamesChangedListener.java
@@ -1,6 +1,6 @@
 package com.ss20.se2.monopoly.network;
 
-import com.ss20.se2.monopoly.pojo.LocallyFoundGame;
+import com.ss20.se2.monopoly.models.LocallyFoundGame;
 
 import java.util.List;
