--- app/src/main/java/com/ss20/se2/monopoly/pojo/models/GamePiece.java
@@ -1,14 +1,0 @@
-package com.ss20.se2.monopoly.pojo.models;
-
-public class GamePiece{
-
-	private String name;
-
-	public GamePiece(String name){
-		this.name = name;
-	}
-
-	public String getName(){
-		return name;
-	}
-}
